The desired control mode can now be selected and published
There are now the following control modes: Remote Control, Follow Wall, Roundtrip with obstacles, Lane Detection, Exploration.
It is now possible to select the desired control mode in the drop-down list. In addition to that, the selected mode is published as a String over the topic /pses_basis/mode_control